【转矩逆变器Only】无感成功，就是目前的控制系数下电机加速有点慢，25A可以上到90 elec.rad/s。注意实验和仿真转速符号有不同：：#define ELECTRICAL_SPEED_FEEDBACK    (-esoaf.xOmg) // 薄片电机实验正iq产生负转速

+++ The acmsimcv5/pmsm_observer.h
@@ -159,7 +159,7 @@ struct ObserverForExperiment{
         REAL x[NS_CHEN_2021];
         REAL ell[NS_CHEN_2021];
 
-        REAL bool_ramp_load_torque; // TRUE for 4th order ESO
+        int bool_ramp_load_torque; // TRUE for 4th order ESO
         REAL omega_ob; // one parameter tuning
         REAL set_omega_ob;
